openvas-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-gvm==26.0.0
lxml==5.3.0
httpx==0.27.2
pytest==8.3.5

google-auth==2.29.0
google-api-python-client==2.128.0